relocated _validate_assets_sbml_ids to sbml_dfs_utils since it is useful outside of the network ops
